requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.0
twilio==8.10.0
boto3==1.34.0
firebase-admin==6.4.0
